% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RecMed}
\alias{RecMed}
\title{Mediation Analysis With Recuurent event outcome}
\usage{
RecMed(
  data,
  exposure,
  mediator,
  pretreat_confounder = NULL,
  posttreat_confounder = NULL,
  mediator_dist = "normal",
  B_number,
  B_points,
  B_seed = 123,
  g_computation = FALSE,
  estimation = TRUE
)
}
\arguments{
\item{data}{data.frame.}

\item{exposure}{character.}

\item{mediator}{character.}

\item{pretreat_confounder}{character vector.}

\item{posttreat_confounder}{character vector.}

\item{mediator_dist}{character.}

\item{B_number}{integer.}

\item{B_points}{numeric vector.}

\item{B_seed}{integer.}

\item{g_computation}{logical.}

\item{estimation}{logical.}
}
\description{
Perform causal mediation analysis with recurrent event outcomes, using either binary or continuous mediator. The distribution of the mediator is estimated through logistic or linear regression models. The distribution of the exposure and outcome are estimated by logistic regression and proportional mean model, respectively. This estimation method requires only two of the three models - exposure, mediator, and outcome - to be correctly specified in order to obtain an unbiased estimate.
}
\examples{
\dontrun{
RecMed(data = sim_data1, exposure = "A", mediator = "M", pretreat_confounder = c("C1", "C2"),
 B_number = 1000 , B_points = 24 * c(0.2,0.4,0.5,0.6,0.8))
}

}
\seealso{
...
}
