% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RecMed}
\alias{RecMed}
\title{Mediation Analysis With Recurrent event outcome}
\usage{
RecMed(
  data,
  exposure,
  mediator,
  pretreat_confounder = NULL,
  posttreat_confounder = NULL,
  mediator_dist = "normal",
  B_number,
  B_points,
  B_seed = 123,
  g_computation = FALSE,
  estimation = TRUE
)
}
\arguments{
\item{data}{a data.frame, containing the variables in the model. Specifically, column names "id", "t.start", "t.stop", and "event" are required.}

\item{exposure}{a character value specifying the column name of exposure.}

\item{mediator}{a character value specifying the column name of mediator.}

\item{pretreat_confounder}{character vector specifying the column names of pre-treatment confounders.}

\item{posttreat_confounder}{character vector specifying the column names of post-treatment confounders.}

\item{mediator_dist}{character. If "normal", linear regression is used to predict mediator. If "bernoulli", logistic regression is used to predict mediator.}

\item{B_number}{integer; the times of bootstrapping.}

\item{B_points}{a numeric vector; the time points that used to compute pointwise 95 percentage confidence intervals by bootstrapping.}

\item{B_seed}{integer. Setting seed.}

\item{g_computation}{a optional logical value. The default is FALSE. If TRUE, only the model of exposure and mediator will be used, but then there is no triple robustness.}

\item{estimation}{logical.}
}
\description{
Perform causal mediation analysis with recurrent event outcomes, using either binary or continuous mediator. The distribution of the mediator is estimated through logistic or linear regression models. The distribution of the exposure and outcome are estimated by logistic regression and proportional mean model, respectively. This estimation method requires only two of the three models - exposure, mediator, and outcome - to be correctly specified in order to obtain an unbiased estimate.
}
\examples{
\dontrun{
RecMed(data = sim_data1, exposure = "A", mediator = "M", pretreat_confounder = c("C1", "C2"),
 B_number = 1000 , B_points = 24 * c(0.2,0.4,0.5,0.6,0.8))
}
}
\seealso{
...
}
